Remove `halt_callback_chains_on_return_false` config from test app
ref: https://github.com/cookpad/kuroko2/pull/112

This option is no longer available on Rails 5.2 or newer.
It changes Active Record's callbacks behavior if a callback returns
false. But I've confirmed the all AR callbacks in this project doesn't
return false.

So we can remove this option safely.

By the way, we use kuroko2 with `halt_callback_chains_on_return_false = false`
over 3 years, and we don't have any problem related to this option.